+++ src/bot.js
@@ -7,7 +7,7 @@ import "dayjs/locale/fr.js";
 dayjs.extend(RelativeTime);
 dayjs.locale("fr");
 
-import { Client, GatewayIntentBits, Partials } from "discord.js";
+import { Client, Events, GatewayIntentBits, Partials } from "discord.js";
 import { Low } from "lowdb";
 import { JSONFile } from "lowdb/node";
 
@@ -90,7 +90,7 @@ const client = new Client({
 client.db = db; // db cache
 
 // Create bot startup
-client.once("ready", async () => {
+client.once(Events.ClientReady, async () => {
   // Time variables
   const tomorrow = dayjs().add(1, "day").hour(8).minute(0).second(0);
   const frequency = 24 * 60 * 60 * 1000; // 24 hours in ms
@@ -116,33 +116,33 @@ client.once("ready", async () => {
 });
 
 // Create an event listener for messages
-client.on("messageCreate", onMessageCreate);
-client.on("messageReactionAdd", onReactionAdd);
+client.on(Events.MessageCreate, onMessageCreate);
+client.on(Events.MessageReactionAdd, onReactionAdd);
 
 // listener for buttons/modals
-client.on("interactionCreate", onInteractionCreate);
+client.on(Events.InteractionCreate, onInteractionCreate);
 
 // listeners for LOGS
-client.on("messageDelete", onMessageDelete);
-client.on("messageUpdate", onMessageUpdate);
+client.on(Events.MessageDelete, onMessageDelete);
+client.on(Events.MessageUpdate, onMessageUpdate);
 
-client.on("roleCreate", onRoleCreate);
-client.on("roleDelete", onRoleDelete);
-client.on("roleUpdate", onRoleUpdate);
+client.on(Events.GuildRoleCreate, onRoleCreate);
+client.on(Events.GuildRoleDelete, onRoleDelete);
+client.on(Events.GuildRoleUpdate, onRoleUpdate);
 
-client.on("channelCreate", onChannelCreate);
-client.on("channelDelete", onChannelDelete);
-client.on("channelUpdate", onChannelUpdate);
+client.on(Events.ChannelCreate, onChannelCreate);
+client.on(Events.ChannelDelete, onChannelDelete);
+client.on(Events.ChannelUpdate, onChannelUpdate);
 
-client.on("threadCreate", onThreadCreate);
-client.on("threadDelete", onThreadDelete);
+client.on(Events.ThreadCreate, onThreadCreate);
+client.on(Events.ThreadDelete, onThreadDelete);
 
-client.on("guildBanAdd", onGuildBanAdd);
-client.on("guildBanRemove", onGuildBanRemove);
+client.on(Events.GuildBanAdd, onGuildBanAdd);
+client.on(Events.GuildBanRemove, onGuildBanRemove);
 
-client.on("guildMemberAdd", onGuildMemberAdd);
-client.on("guildMemberRemove", onGuildMemberRemove);
-client.on("guildMemberUpdate", onGuildMemberUpdate);
+client.on(Events.GuildMemberAdd, onGuildMemberAdd);
+client.on(Events.GuildMemberRemove, onGuildMemberRemove);
+client.on(Events.GuildMemberUpdate, onGuildMemberUpdate);
 
 // Log the bot in
 client.login(process.env.TOKEN);